examples: rename example operator
Rename the operator to match the pattern
<solution-name>-operator:
example-operator -> example-solution-operator
