+++ examples/metalk8s-solution-example/operator/.gitignore
