+++ examples/metalk8s-solution-example/operator/deploy/crds/examplesolution_v1alpha1_versionserver_cr.yaml
