+++ examples/metalk8s-solution-example/operator/deploy/operator.yaml
@@ -1,24 +1,24 @@
 apiVersion: apps/v1
 kind: Deployment
 metadata:
-  name: example-operator
+  name: example-solution-operator
 spec:
   replicas: 1
   selector:
     matchLabels:
-      name: example-operator
+      name: example-solution-operator
   template:
     metadata:
       labels:
-        name: example-operator
+        name: example-solution-operator
     spec:
-      serviceAccountName: example-operator
+      serviceAccountName: example-solution-operator
       containers:
-        - name: example-operator
+        - name: example-solution-operator
           # Replace this with the built image name
           image: REPLACE_IMAGE
           command:
-          - example-operator
+          - example-solution-operator
           imagePullPolicy: Always
           env:
             - name: WATCH_NAMESPACE
@@ -30,4 +30,4 @@ spec:
                 fieldRef:
                   fieldPath: metadata.name
             - name: OPERATOR_NAME
-              value: "example-operator"
+              value: "example-solution-operator"
