+++ examples/metalk8s-solution-example/operator/deploy/role_binding.yaml
@@ -1,11 +1,11 @@
 kind: RoleBinding
 apiVersion: rbac.authorization.k8s.io/v1
 metadata:
-  name: example-operator
+  name: example-solution-operator
 subjects:
 - kind: ServiceAccount
-  name: example-operator
+  name: example-solution-operator
 roleRef:
   kind: Role
-  name: example-operator
+  name: example-solution-operator
   apiGroup: rbac.authorization.k8s.io
