+++ examples/metalk8s-solution-example/operator/deploy/service_account.yaml
@@ -1,4 +1,4 @@
 apiVersion: v1
 kind: ServiceAccount
 metadata:
-  name: example-operator
+  name: example-solution-operator
