+++ examples/metalk8s-solution-example/operator/pkg/apis/apis.go
