+++ examples/metalk8s-solution-example/operator/pkg/apis/examplesolution/group.go
