+++ examples/metalk8s-solution-example/operator/pkg/apis/examplesolution/v1alpha1/clockserver_types.go
