+++ examples/metalk8s-solution-example/operator/pkg/apis/examplesolution/v1alpha1/versionserver_types.go
