+++ examples/metalk8s-solution-example/operator/pkg/controller/controller.go
