+++ examples/metalk8s-solution-example/operator/version/version.go
