+++ examples/metalk8s-solution-example/ui/src/constants.js
@@ -6,7 +6,7 @@ export const LABEL_PART_OF = 'app.kubernetes.io/part-of';
 export const LABEL_VERSION = 'app.kubernetes.io/version';
 export const SOLUTION_NAME = 'example-solution';
 export const SOLUTION_API_GROUP = `${SOLUTION_NAME}.metalk8s.scality.com`;
-export const DEPLOYMENT_NAME = 'example-operator';
+export const DEPLOYMENT_NAME = 'example-solution-operator';
 export const OPERATOR_NAME = 'example-solution-operator';
 export const SOLUTION_CONFIGMAP_NAME = 'metalk8s-solutions';
 export const SOLUTION_CONFIGMAP_NAMESPACE = 'metalk8s-solutions';
